unify code
